One way of reorganizing our file structure.
Put all components belonging only to one page to that page's component
folder. Components that are and can be used by more than one page and/or
components from a different page should be put in the top-level components
folder.
