+++ src/pages/CalendarPage/components/EventCard/index.js
