+++ src/pages/CalendarPage/components/EventCard/styles.js
