+++ src/pages/CalendarPage/index.js
@@ -3,9 +3,9 @@ import React from 'react';
 import { withStyles } from '@material-ui/core/styles';
 import { Grid, Paper, Container, Button } from '@material-ui/core';
 
-import Calendar from '../../components/Calendar';
-import EventCard from '../../components/EventCard';
-import EventList from '../../components/EventList';
+import Calendar from './components/Calendar';
+import EventCard from './components/EventCard';
+import EventList from './components/EventList';
 import { getAllEvents } from '../../services/events';
 
 import styles from './styles';
