+++ src/pages/EventDetailsPage/components/DeleteEditButtons/styles.js
