+++ src/pages/EventDetailsPage/components/EventDetails/styles.js
