+++ src/pages/EventEditPage/components/ChipListInput/styles.js
