+++ src/pages/EventEditPage/components/EventEditForm/form_chip_list_input.js
