+++ src/pages/EventEditPage/components/EventEditForm/form_multi_chip_select.js
