+++ src/pages/EventEditPage/components/MultiChipSelect/index.js
