+++ src/pages/EventEditPage/event_edit.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import moment from 'moment';
 
 import { getEventById, setEventDetails } from '../../services/events';
-import EventEditForm from '../../components/EventEditForm';
+import EventEditForm from './components/EventEditForm';
 
 class EventEditPage extends React.Component {
   constructor(props) {
