+++ src/pages/SignUpPage/components/SignUpForm/schema.js
